Remove: remove parseConfig -> parseUtils
