+++ Makefile
@@ -4,8 +4,7 @@ SFILE = Server.cpp \
 		debug.cpp \
 		Socket.cpp \
 		ServerSocket.cpp \
-		./utils/parseUtils.cpp \
-		parseConfig.cpp
+		./utils/parseUtils.cpp 
 		
 
 # SOBJ = $(SFILE:.c=.o)
